# api.yaml
openapi: 3.0.2

info:
  version: 0.0.1
  title: Contact Api
  description: Contact Api for Open Web Technology Challenge

servers:
  -   url: "localhost:8080"
      description: Contact Api

tags:
  - name: "Contact"
  - name: "Skill"

paths:

  /contact:
    get:
      tags:
        - "Contact"
      description: get all contacts
      operationId: getAllContacts
      responses:
        '200':
          description: Return all the contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      tags:
        - "Contact"
      description: Create contact
      operationId: createContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contact"
      responses:
        '201':
          description: Return new contact model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'

  /contact/{contactId}:
    get:
      tags:
        - "Contact"
      description: Get one contact
      operationId: getOneContact
      parameters:
        - in: path
          name: contactId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Return contact model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
    put:
      tags:
        - "Contact"
      description: Update one contact
      operationId: updateContact
      parameters:
        - in: path
          name: contactId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contact"
      responses:
        '200':
          description: Return updated contact model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
    delete:
      tags:
        - "Contact"
      description: Delete one contact
      operationId: deleteContact
      parameters:
        - in: path
          name: contactId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Contact is deleted

  /contact/{contactId}/skill:
    get:
      tags:
        - "Contact"
      description: Get all skills of one contact
      operationId: getAllSkillFromContact
      parameters:
        - in: path
          name: contactId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Return list of skill
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: "#/components/schemas/Skill"

  /contact/{contactId}/skill/{skillId}:
    put:
      tags:
        - "Contact"
      description: Add a skill to a contact
      operationId: addSkill
      parameters:
        - in: path
          name: contactId
          required: true
          schema:
            type: string
        - in: path
          name: skillId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Return added skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      tags:
        - "Contact"
      description: Remove a skill from a contact
      operationId: removeSkill
      parameters:
        - in: path
          name: contactId
          required: true
          schema:
            type: string
        - in: path
          name: skillId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Skill was correctly removed from contact

  /skill:
    get:
      tags:
        - "Skill"
      description: get all the skills
      operationId: getAllSkill
      responses:
        '200':
          description: Return all the skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
    post:
      tags:
        - "Skill"
      description: Create skill
      operationId: createSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Skill"
      responses:
        '201':
          description: Return new skill model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'

  /skill/{skillId}:
    get:
      tags:
        - "Skill"
      description: Get one skill
      operationId: getOneSkill
      parameters:
        - in: path
          name: skillId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Return skill model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
    put:
      tags:
        - "Skill"
      description: Update one skill
      operationId: updateSkill
      parameters:
        - in: path
          name: skillId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Skill"
      responses:
        '200':
          description: Return updated skill model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
    delete:
      tags:
        - "Skill"
      description: Delete one skill
      operationId: deleteSkill
      parameters:
        - in: path
          name: skillId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Skill is deleted

components:
  schemas:
    Skill:
      description: Skill model
      type: object
      properties:
        id:
          description: Id of the skill.
          type: string
          readOnly: true
        name:
          description: Skill name
          type: string
        level:
          $ref: "#/components/schemas/Level"

    Contact:
      description: Contact model
      type: object
      properties:
        id:
          description: Id of the contact.
          type: string
          readOnly: true
        firstname:
          description: First name of the contact
          type: string
        lastname:
          description: Last name of the contact
          type: string
        fullname:
          description: Full name of the contact
          type: string
          readOnly: true
        address:
          description: Address of the contact
          type: string
        email:
          description: Email of the contact
          type: string
          format: email
        mobile:
          description: Mobile phone number of the contact
          type: string

    Level:
      description: Skill level as enum
      type: string
      enum: ['LOW', 'INTERMEDIATE', 'HIGH']